% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popular.R
\name{popular}
\alias{popular}
\title{popular(url, type="likes")}
\usage{
popular(url, type = "likes")
}
\arguments{
\item{url}{An object that used the `web_page()` function.}

\item{type}{"likes" returns response with the most likes. "author" returns response confirmed by author of question.}
}
\value{
reponse Returns a response. This will be of a character type.
}
\description{
Loads Returns most popular response on webpage.
}
\examples{
#'popular('https://stackoverflow.com/questions/3505701/grouping-functions-tapply-by-aggregate-and-the-apply-family', type="likes")
}
