% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_stats.R
\name{response_stats}
\alias{response_stats}
\title{response_stats(url)}
\usage{
response_stats(url)
}
\arguments{
\item{url}{An object that used the `web_page()` function.}
}
\value{
Returns a vector of length 4 that includes:

n Returns number of responses.

m Average reputation score.

t Top reputation score.

a Author reputation score.
}
\description{
Returns general statistics about question being asked on a Stack Overflow page.
}
\examples{
#'response_stats('https://stackoverflow.com/questions/3505701/grouping-functions-tapply-by-aggregate-and-the-apply-family')
}
