% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_page.R
\name{web_page}
\alias{web_page}
\title{web_page(url)}
\usage{
web_page(url)
}
\arguments{
\item{url}{Full URL to stack overflow response webpage. Character type.}
}
\value{
q The question being asked on the given Stack Overflow response webpage. Character type. If unsuccessful corresponding error will be returned.
}
\description{
Loads Stack Overflow response webpage.
}
\examples{
#'web_page('https://stackoverflow.com/questions/3505701/grouping-functions-tapply-by-aggregate-and-the-apply-family')
}
